Cleanup scipt loading, async safety, queue id
