+++ src/main/java/com/denizenscript/clientizen/Clientizen.java
@@ -2,9 +2,9 @@ package com.denizenscript.clientizen;
 
 import com.denizenscript.clientizen.debuggui.ClientizenDebugScreen;
 import com.denizenscript.clientizen.events.ClientizenScriptEventRegistry;
-import com.denizenscript.clientizen.network.ClientRunHandler;
 import com.denizenscript.clientizen.network.NetworkManager;
 import com.denizenscript.clientizen.objects.ClientizenObjectRegistry;
+import com.denizenscript.clientizen.scripts.ClientScriptHelper;
 import com.denizenscript.clientizen.scripts.commands.ClientizenCommandRegistry;
 import com.denizenscript.clientizen.scripts.containers.ClientizenContainerRegistry;
 import com.denizenscript.clientizen.tags.ClientizenTagContext;
@@ -75,7 +75,7 @@ public class Clientizen implements ClientModInitializer {
 		// Initialize Clientizen systems
 		NetworkManager.init();
 		ClientizenDebugScreen.register();
-		ClientRunHandler.init();
+		ClientScriptHelper.init();
 
 		// Check for the client scripts folder
 		File scriptsFolder = DenizenCore.implementation.getScriptFolder();
